requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
diff-match-patch==20241021
Django==5.2.11
django-cors-headers==4.9.0
django-import-export==4.4.0
django-ninja==1.5.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.5
tablib==3.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==202602
user-agents==2.2.0
psycopg2-binary==2.9.11
gunicorn==25.0.3
whitenoise==6.11.0


